+++ R/get_pvol_cz.R
@@ -6,7 +6,10 @@ get_pvol_cz <- function(radar, time, ..., call = rlang::caller_env()) {
   # Here all urls are generated
   params <- c("z", "u", "v", "w", "zdr", "rhohv", "phidp")
   urls <- glue::glue(
-    "http://opendata.chmi.cz/meteorology/weather/radar/sites/{substr(radar,3,5)}/vol_{params}/hdf5/"
+    getOption(
+      "getRad.cz_url_format",
+      default = "http://opendata.chmi.cz/meteorology/weather/radar/sites/{substr(radar,3,5)}/vol_{params}/hdf5/"
+    )
   )
   rlang::check_installed(
     c("lubridate", "tidyr", "xml2", "rhdf5"),